% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateCI_B}
\alias{calculateCI_B}
\title{calculateCI_B}
\usage{
calculateCI_B(
  X,
  Y,
  index,
  alpha = 0.05,
  known = F,
  sigma = NULL,
  mse = NULL,
  inverseXT_X = NULL,
  B = NULL
)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{index}{the index of Beta}

\item{alpha}{(optional) the value of significance level}

\item{known}{(optional) a boolean parameter that takes (T/F) as value}

\item{sigma}{(optional) the standard error of the dataset and it has  default value (NULL)}

\item{mse}{(optional) mean square error}

\item{inverseXT_X}{(optional) inverse of x-tranposed multiplied by x}

\item{B}{(optional) vector of estimated regression coefficients.}
}
\value{
the confidence interval of Beta (string)
}
\description{
This function calculates the confidence interval of any Beta estimator in a multiple regression model
}
