% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateDFE}
\alias{calculateDFE}
\title{Compute Residual degree of freedom}
\usage{
calculateDFE(X)
}
\arguments{
\item{X}{X matrix constructed with constructX()}
}
\value{
residual degree of freedom
}
\description{
Compute Residual degree of freedom
}
