% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateDFR}
\alias{calculateDFR}
\title{Compute Regression degree of freedom}
\usage{
calculateDFR(X)
}
\arguments{
\item{X}{X matrix constructed with constructX()}
}
\value{
regression degree of freedom
}
\description{
Compute Regression degree of freedom
}
