% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateDFT}
\alias{calculateDFT}
\title{Compute Total degree of freedom}
\usage{
calculateDFT(X)
}
\arguments{
\item{X}{X matrix constructed with constructX()}
}
\value{
total degree of freedom
}
\description{
Compute Total degree of freedom
}
