% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateFVALUE}
\alias{calculateFVALUE}
\title{Calculate F Value}
\usage{
calculateFVALUE(X, Y, msr = NULL, mse = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{msr}{(optional) mean square regression}

\item{mse}{(optional) mean square error}
}
\value{
anova f value
}
\description{
Calculate F Value
}
