% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateMSE}
\alias{calculateMSE}
\title{Compute the Mean Square Error}
\usage{
calculateMSE(X, Y, sse = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{sse}{(optional) residual sum of squares}
}
\value{
mean square error
}
\description{
Compute the Mean Square Error
}
