% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateMSR}
\alias{calculateMSR}
\title{Compute the Mean Square Regression}
\usage{
calculateMSR(X, Y, ssr = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{ssr}{(optional) regression sum of squares}
}
\value{
mean square regression
}
\description{
Compute the Mean Square Regression
}
