% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateRSQUARED}
\alias{calculateRSQUARED}
\title{Calculate R-Squared value}
\usage{
calculateRSQUARED(X, Y, SSR = NULL, SST = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{SSR}{regression sum of squares}

\item{SST}{total sum of squares}
}
\value{
r-squared
}
\description{
Calculate R-Squared value
}
