% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{calculateRSQUAREDADJUSTED}
\alias{calculateRSQUAREDADJUSTED}
\title{Calculate the Adjusted R-Squared value}
\usage{
calculateRSQUAREDADJUSTED(X, Y, RSQUARED = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{RSQUARED}{rsquared, calculated with calculateRSQUARED()}
}
\description{
Calculate the Adjusted R-Squared value
}
