% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{coerelation_matrix}
\alias{coerelation_matrix}
\title{Calculate the correlation matrix}
\usage{
coerelation_matrix(X)
}
\arguments{
\item{X}{X matrix constructed with constructX()}
}
\value{
correlation matrix
}
\description{
Calculate the correlation matrix
}
