% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{compute_y0}
\alias{compute_y0}
\title{Get the mean Y value at certain X values}
\usage{
compute_y0(X, Y, X0, B = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{X0}{X values}

\item{B}{(optional) vector of estimated regression coefficients.}
}
\value{
y0
}
\description{
Get the mean Y value at certain X values
}
