% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{compute_y_pred}
\alias{compute_y_pred}
\title{Compute Predicted Y values}
\usage{
compute_y_pred(X, Y, B = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{B}{(optional) vector of estimated regression coefficients.}
}
\value{
y-predicted
}
\description{
Compute Predicted Y values
}
