% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{computeB}
\alias{computeB}
\title{Estimate Regression Coefficients B}
\usage{
computeB(X, Y, Inv_XT_X = NULL, XT_Y = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{Inv_XT_X}{(optional) inverse of x-tranposed multiplied by x}

\item{XT_Y}{(optional) x-tranposed multiplied by y}
}
\value{
B vector
}
\description{
Estimate Regression Coefficients B
}
