% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{computeCI_new_observation}
\alias{computeCI_new_observation}
\title{Confidence Interval for a New Observation}
\usage{
computeCI_new_observation(
  X,
  Y,
  X0,
  alpha = 0.05,
  sigma = NULL,
  mse = NULL,
  known = F,
  B = NULL,
  inverse_XTX = NULL
)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{alpha}{(optional) the value of significance level}

\item{sigma}{(optional) the standard error of the dataset and it has  default value (NULL)}

\item{mse}{(optional) mean square error}

\item{known}{(optional) a boolean parameter that takes (T/F) as value}

\item{B}{(optional) vector of estimated regression coefficients.}

\item{inverse_XTX}{(optional) inverse of x-tranposed multiplied by x}
}
\value{
new observation confidence interval (string)
}
\description{
Confidence Interval for a New Observation
}
