% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{computeSSR}
\alias{computeSSR}
\title{Compute Regression Sum of Squares}
\usage{
computeSSR(X, Y, B = NULL, Inv_XT_X = NULL, XT_Y = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{B}{(optional) vector of estimated regression coefficients. \link{The remaining parameters are not needed if B is provided}}

\item{Inv_XT_X}{(optional) inverse of x-tranposed multiplied by x}

\item{XT_Y}{(optional) x-tranposed multiplied by y}
}
\value{
regression sum of squares
}
\description{
Compute Regression Sum of Squares
}
