% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{computeSST}
\alias{computeSST}
\title{Compute Total Sum of Squares}
\usage{
computeSST(X, Y, YT_Y = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{YT_Y}{transpose of Y multiplied by Y}
}
\value{
total sum of squares
}
\description{
Compute Total Sum of Squares
}
