% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{constructinverseXT_X}
\alias{constructinverseXT_X}
\title{Construct The Inverse of X-transpose X}
\usage{
constructinverseXT_X(X, XT_X = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{XT_X}{(optional) x-tranposed multiplied by x}
}
\value{
the inverse of the transpose of X multiplied by X
}
\description{
Construct The Inverse of X-transpose X
}
