% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{constructX}
\alias{constructX}
\title{Construct X matrix}
\usage{
constructX(dfX)
}
\arguments{
\item{dfX}{dataframe with selected X features only. Can be created with new_dataframe_x().}
}
\value{
X matrix, first column contains 1s, other rows contain the X features in dfX
}
\description{
Construct X matrix
}
