% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{constructXT_X}
\alias{constructXT_X}
\title{Construct X-transpose X}
\usage{
constructXT_X(X)
}
\arguments{
\item{X}{X matrix constructed with constructX()}
}
\value{
transpose of X multiplied by X
}
\description{
Construct X-transpose X
}
