% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{constructXT_Y}
\alias{constructXT_Y}
\title{Construct X-transpose Y}
\usage{
constructXT_Y(X, Y)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}
}
\value{
transpose of X multiplied by Y
}
\description{
Construct X-transpose Y
}
