% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{constructY}
\alias{constructY}
\title{Construct Y matrix}
\usage{
constructY(dfY)
}
\arguments{
\item{dfY}{dataframe with 'y' column. Can be created with new_dataframe_y().}
}
\value{
Y values in vector
}
\description{
Construct Y matrix
}
