% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{constructYT_Y}
\alias{constructYT_Y}
\title{Construct Y-transpose Y}
\usage{
constructYT_Y(Y)
}
\arguments{
\item{Y}{Y vector constructed with constructY()}
}
\value{
transpose of Y multiplied by Y
}
\description{
Construct Y-transpose Y
}
