% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{get_ypred_values}
\alias{get_ypred_values}
\title{Get specific Y value}
\usage{
get_ypred_values(X, Y, n, y_pred = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{n}{index of Y to get}

\item{y_pred}{(optional) y-predicted, calculated with compute_y_pred()}
}
\value{
specific Y value
}
\description{
Get specific Y value
}
