% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{getBvalues}
\alias{getBvalues}
\title{Get a specific B value}
\usage{
getBvalues(X, Y, n, B = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{n}{index of B to get}

\item{B}{(optional) vector of estimated regression coefficients.}
}
\value{
specific B value
}
\description{
Get a specific B value
}
