% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{new_dataframe_x}
\alias{new_dataframe_x}
\title{Create dataframe 'dfX' with selected features}
\usage{
new_dataframe_x(df, ...)
}
\arguments{
\item{...}{names of selected features columns}
}
\value{
dataframe dfX
}
\description{
Create dataframe 'dfX' with selected features
}
