% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{new_dataframe_y}
\alias{new_dataframe_y}
\title{Create dataframe 'dfY' with Y values}
\usage{
new_dataframe_y(new_vector)
}
\arguments{
\item{new_vector}{vector with y values}
}
\value{
dataframe dfY
}
\description{
Create dataframe 'dfY' with Y values
}
