% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{partialF}
\alias{partialF}
\title{Calculate the partial F value}
\usage{
partialF(X, Y, idx_vec, SSRT = NULL, MSEM = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{idx_vec}{vector with the indices of reduced model (non-tested features)}

\item{SSRT}{total regression sum of squares}

\item{MSEM}{model's mean square error}
}
\value{
partial f-value
}
\description{
Calculate the partial F value
}
