% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{regular_error}
\alias{regular_error}
\title{Compute the Regular Error}
\usage{
regular_error(X, Y, B = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{B}{(optional) vector of estimated regression coefficients.}
}
\value{
regular error vector
}
\description{
Compute the Regular Error
}
