% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{standard_error}
\alias{standard_error}
\title{Compute the Standard Error}
\usage{
standard_error(X, Y, E = NULL, B = NULL, MSE = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{E}{(optional) regular error}

\item{B}{(optional) vector of estimated regression coefficients}

\item{MSE}{(optional) mean square error}
}
\value{
standard error matrix
}
\description{
Compute the Standard Error
}
