% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{student_error}
\alias{student_error}
\title{Compute the Student Error}
\usage{
student_error(X, Y, E = NULL, B = NULL, MSE = NULL, inv_XT_X = NULL)
}
\arguments{
\item{X}{X matrix constructed with constructX()}

\item{Y}{Y vector constructed with constructY()}

\item{E}{(optional) regular error}

\item{B}{(optional) vector of estimated regression coefficients}

\item{MSE}{(optional) mean square error}

\item{inv_XT_X}{(optional) inverse of x-tranposed multiplied by x}
}
\value{
student error matrix
}
\description{
Compute the Student Error
}
