% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{VIF}
\alias{VIF}
\title{Calculate the variance inflation factor}
\usage{
VIF(X)
}
\arguments{
\item{X}{X matrix constructed with constructX()}
}
\value{
vif
}
\description{
Calculate the variance inflation factor
}
